+++ packages/docs-article-templates/src/helper/user-settings.ts
@@ -6,4 +6,4 @@ import { workspace } from 'vscode';
 export const gitHubID: string = workspace.getConfiguration('docs.templates').githubid;
 export const alias: string = workspace.getConfiguration('docs.templates').alias;
 export const missingValue: string = 'NO VALUE SET';
-export const learnRepoId: string = workspace.getConfiguration('docs.templates').learn_repo_id;
+export const templateRepo: string = workspace.getConfiguration('docs.templates').template_repo;