+++ packages/docs-markdown-to-mail/src/helper/common.ts
@@ -107,3 +107,20 @@ export function tryFindFile(rootPath: string, fileName: string) {
 	postWarning(`Unable to find a file named "${fileName}", recursively at root "${rootPath}".`);
 	return undefined;
 }
+
+export function showStatusMessage(message: string) {
+	const { msTimeValue } = generateTimestamp();
+	output.appendLine(`[${msTimeValue}] - ${message}`);
+}
+
+export function getCommunicationDate() {
+	const date = new Date(Date.now());
+	const month = date.toLocaleString('default', { month: 'long' });
+	const year = date.getFullYear();
+	const communicationDate = `${month} ${year}`;
+	return communicationDate;
+}
+
+export function updateSiteRelativeLinks(content: string) {
+	return content.replace(/\(\//, '(https://review.docs.microsoft.com/');
+}