+++ packages/docs-markdown/src/helper/metadata.ts
@@ -1,6 +1,10 @@
 import { updateMetadataDate } from '../controllers/metadata-controller';
-import { workspace, window } from 'vscode';
-import { noActiveEditorMessage, toShortDate, isMarkdownFileCheck } from './common';
+import { workspace, window, Position, Range } from 'vscode';
+import {
+	noActiveEditorMessage,
+	toShortDate,
+	isMarkdownYamlFileCheckWithoutNotification
+} from './common';
 import { findReplacement } from './utility';
 
 const blacklist: string[] = [];
@@ -19,7 +23,7 @@ export async function metadataDateReminder() {
 		return;
 	}
 
-	if (!isMarkdownFileCheck(editor, false)) {
+	if (!isMarkdownYamlFileCheckWithoutNotification(editor)) {
 		return;
 	}
 
@@ -44,7 +48,7 @@ export async function metadataDateReminder() {
 			`ms.date: ${toShortDate(new Date())}`,
 			msDateRegex
 		);
-		if (msDateReplacement) {
+		if (msDateReplacement && checkIfShortDate(msDateReplacement, editor)) {
 			const syncDate = await window.showInformationMessage(
 				"Would you like to update ms.date to today's date?",
 				'Update'
@@ -54,6 +58,37 @@ export async function metadataDateReminder() {
 			}
 		}
 	}
+}
+
+function checkIfShortDate(msDateReplacement, editor) {
+	const shortDay = new Date().getDate();
+	const shortMonth = new Date().getMonth() + 1;
+	const year = new Date().getFullYear();
+	const month = shortMonth > 9 ? shortMonth : `0${shortMonth}`;
+	const day = shortDay > 9 ? shortDay : `0${shortDay}`;
+	const shortDayLongMonthDate = `${month}/${shortDay}/${year}`;
+	const shortMonthDayDate = `${shortMonth}/${shortDay}/${year}`;
+	const shortMonthDate = `${shortMonth}/${day}/${year}`;
+	const selectedText = editor.document.getText(
+		new Range(
+			new Position(
+				msDateReplacement.selection.start.line,
+				msDateReplacement.selection.start.character
+			),
+			new Position(msDateReplacement.selection.end.line, msDateReplacement.selection.end.character)
+		)
+	);
+	const msDateShortDayDate = `ms.date: ${shortDayLongMonthDate}`;
+	const msDateShortMonthDayDate = `ms.date: ${shortMonthDayDate}`;
+	const msDateShortMonthDate = `ms.date: ${shortMonthDate}`;
+	if (
+		selectedText === msDateShortMonthDate ||
+		selectedText === msDateShortMonthDayDate ||
+		selectedText === msDateShortDayDate
+	) {
+		return false;
+	}
+	return true;
 }
 
 export async function disableMetadataDateReminder() {