+++ packages/docs-markdown/src/test/suite/helper/metadata.test.ts
@@ -9,6 +9,7 @@ import * as telemetry from '../../../helper/telemetry';
 import * as utility from '../../../helper/utility';
 import { loadDocumentAndGetItReady } from '../../test.common/common';
 import { resolve } from 'path';
+import sinon = require('sinon');
 
 chai.use(spies);
 
@@ -16,7 +17,6 @@ const expect = chai.expect;
 
 suite('Metadata Helper', () => {
 	// eslint-disable-next-line @typescript-eslint/no-var-requires
-	const sinon = require('sinon');
 
 	setup(async () => {
 		sinon.stub(workspace, 'onWillSaveTextDocument');
@@ -50,12 +50,20 @@ suite('Metadata Helper', () => {
 		expect(spy).to.have.been.called();
 	});
 	test('metadataDateReminder => user selects to update ms.date', async () => {
+		sinon.stub(workspace, 'getConfiguration').returns({
+			get: () => true,
+			has: () => true,
+			inspect: () => {
+				return { key: '' };
+			},
+			update: () => Promise.resolve()
+		});
 		const stub = sinon.stub(metadata, 'updateMetadataDate');
 		const config = workspace.getConfiguration('markdown');
 		await config.update('enableMetadataDateReminder', true);
 		// stub user response choosing to update the ms.date
 		const stubShowInformationMessage = sinon.stub(window, 'showInformationMessage');
-		stubShowInformationMessage.onCall(0).resolves(Promise.resolve('Update') as Thenable<any>);
+		stubShowInformationMessage.onCall(0).resolves(Promise.resolve('Update') as any);
 		const filePath = resolve(
 			__dirname,
 			'../../../../../src/test/data/repo/articles/test/metadata2.md'
@@ -67,12 +75,20 @@ suite('Metadata Helper', () => {
 		stubShowInformationMessage.restore();
 	});
 	test('metadataDateReminder => user selects not to update ms.date', async () => {
+		sinon.stub(workspace, 'getConfiguration').returns({
+			get: () => true,
+			has: () => true,
+			inspect: () => {
+				return { key: '' };
+			},
+			update: () => Promise.resolve()
+		});
 		const stub = sinon.stub(metadata, 'updateMetadataDate');
 		const config = workspace.getConfiguration('markdown');
 		await config.update('enableMetadataDateReminder', true);
 		// creates user response choosing not to update the ms.date
 		const stubShowInformationMessage = sinon.stub(window, 'showInformationMessage');
-		stubShowInformationMessage.onCall(0).resolves(Promise.resolve(undefined) as Thenable<any>);
+		stubShowInformationMessage.onCall(0).resolves(Promise.resolve(undefined) as any);
 		const filePath = resolve(
 			__dirname,
 			'../../../../../src/test/data/repo/articles/test/metadata3.md'
@@ -84,11 +100,19 @@ suite('Metadata Helper', () => {
 		stubShowInformationMessage.restore();
 	});
 	test('metadataDateReminder => only remind once', async () => {
+		sinon.stub(workspace, 'getConfiguration').returns({
+			get: () => true,
+			has: () => true,
+			inspect: () => {
+				return { key: '' };
+			},
+			update: () => Promise.resolve()
+		});
 		const stub = sinon.stub(utility, 'findReplacement');
 		const config = workspace.getConfiguration('markdown');
 		await config.update('enableMetadataDateReminder', true);
 		const stubShowInformationMessage = sinon.stub(window, 'showInformationMessage');
-		stubShowInformationMessage.onCall(0).resolves(Promise.resolve(undefined) as Thenable<any>);
+		stubShowInformationMessage.onCall(0).resolves(Promise.resolve(undefined) as any);
 		const filePath = resolve(
 			__dirname,
 			'../../../../../src/test/data/repo/articles/test/metadata4.md'
@@ -103,13 +127,56 @@ suite('Metadata Helper', () => {
 		stubShowInformationMessage.restore();
 	});
 	test('disableMetadataDateReminder()', async () => {
+		sinon.stub(workspace, 'getConfiguration').returns({
+			get: () => false,
+			has: () => true,
+			inspect: () => {
+				return { key: '' };
+			},
+			update: () => Promise.resolve()
+		});
 		await metadataHelper.disableMetadataDateReminder();
 		const config = workspace.getConfiguration('markdown');
 		assert.strictEqual(config.get<boolean>('enableMetadataDateReminder'), false);
 	});
 	test('enableMetadataDateReminder()', async () => {
+		sinon.stub(workspace, 'getConfiguration').returns({
+			get: () => true,
+			has: () => true,
+			inspect: () => {
+				return { key: '' };
+			},
+			update: () => Promise.resolve()
+		});
 		await metadataHelper.enableMetadataDateReminder();
 		const config = workspace.getConfiguration('markdown');
 		assert.strictEqual(config.get<boolean>('enableMetadataDateReminder'), true);
+	});
+	test('metadataDateReminder => doesnt remind on short date', async () => {
+		sinon.stub(workspace, 'getConfiguration').returns({
+			get: () => true,
+			has: () => true,
+			inspect: () => {
+				return { key: '' };
+			},
+			update: () => Promise.resolve()
+		});
+
+		const config = workspace.getConfiguration('markdown');
+		await config.update('enableMetadataDateReminder', true);
+		const stubShowInformationMessage = sinon.stub(window, 'showInformationMessage');
+		stubShowInformationMessage.onCall(0).resolves(Promise.resolve(undefined) as any);
+		const filePath = resolve(
+			__dirname,
+			'../../../../../src/test/data/repo/articles/test/metadata5.md'
+		);
+		await loadDocumentAndGetItReady(filePath);
+		const clock = sinon.useFakeTimers({
+			now: new Date(2021, 1, 2, 0, 0)
+		});
+		await metadataHelper.metadataDateReminder();
+		assert.strictEqual(stubShowInformationMessage.callCount, 0);
+		clock.restore();
+		stubShowInformationMessage.restore();
 	});
 });