+++ packages/docs-markdown/changelog.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 0.2.84 (March 1st, 2021)
+
+- Bug fixes for metadata date nag and add period to alt text.
+
+## 0.2.83 (February 24th, 2021)
+
+- Bug fixes
+- Added feature for Add Docs link by URL to Link quick pick
+- Updated quick-pick-menu ordering
+
 ## 0.2.82 (October 23rd, 2020)
 
 - Update allowlist endpoint